requirements.txt
# Django
Django==3.1.6

# .env 
python-dotenv==0.15.0

# bilder
Pillow==8.1.0

# What you see is what you editor
django-summernote==0.8.11.6

# lint
pylint==2.6.2
pylint-django==2.4.2

#Graphql
graphene==2.1.8
graphene-django==2.15.0

# rest framework
djangorestframework==3.12.2